back float annotaion to budget and revenue
luke asked at the first review to make them int

+++ src/schemas/movies.py
@@ -14,8 +14,8 @@ class MovieBase(BaseModel):
     orig_title: str
     status: str
     orig_lang: str
-    budget: int
-    revenue: int
+    budget: float
+    revenue: float
     country: str
